consolidate apiBaseURL into config and use AuthHeaderPrefix everywhere (#64)
