+++ cmd/context.go
@@ -174,7 +174,7 @@ func runContextList(cmd *cobra.Command, args []string) error {
 		return nil
 	}
 
-	baseURL, err := apiBaseURL(cfg.ServerURL)
+	baseURL, err := config.APIBaseURL(cfg.ServerURL)
 	if err != nil {
 		return fmt.Errorf("parse server URL: %w", err)
 	}
@@ -268,7 +268,7 @@ func runContextUse(cmd *cobra.Command, args []string) error {
 	// the server is unreachable. The downstream best-effort push then
 	// skips its own attempt to avoid emitting a second redundant warning.
 	offlineDetected := false
-	baseURL, parseErr := apiBaseURL(cfg.ServerURL)
+	baseURL, parseErr := config.APIBaseURL(cfg.ServerURL)
 	if cfg.AuthToken != "" && strings.HasPrefix(name, config.TeamContextPrefix) && parseErr == nil {
 		memberships, supported, definitelyNotMember, fetchErr := fetchMembershipsCached(nil, baseURL, cfg.AuthToken)
 		switch {
@@ -352,7 +352,7 @@ func runContextShow(cmd *cobra.Command, args []string) error {
 	//
 	// Compute the REST base URL once and pass it down so stalenessAnnotation
 	// doesn't repeat the work the caller already did.
-	baseURL, baseURLErr := apiBaseURL(cfg.ServerURL)
+	baseURL, baseURLErr := config.APIBaseURL(cfg.ServerURL)
 	if baseURLErr != nil {
 		fmt.Fprintln(cmd.OutOrStdout(), active)
 		return nil
@@ -373,7 +373,7 @@ func runContextShow(cmd *cobra.Command, args []string) error {
 // the slug-keyed memberships endpoint).
 //
 // Callers compute the REST baseURL once and pass it in to avoid two
-// successive apiBaseURL calls.
+// successive config.APIBaseURL calls.
 func stalenessAnnotation(cfg *config.Config, baseURL, active string) (string, bool) {
 	if cfg == nil || cfg.AuthToken == "" {
 		return "", false
