+++ cmd/status.go
@@ -48,7 +48,7 @@ func runStatus(cmd *cobra.Command, args []string) error {
 		return display.AuthError("not authenticated")
 	}
 
-	baseURL, err := apiBaseURL(cfg.ServerURL)
+	baseURL, err := config.APIBaseURL(cfg.ServerURL)
 	if err != nil {
 		return fmt.Errorf("parse server URL: %w", err)
 	}
@@ -57,7 +57,7 @@ func runStatus(cmd *cobra.Command, args []string) error {
 	if err != nil {
 		return fmt.Errorf("create request: %w", err)
 	}
-	req.Header.Set("Authorization", "Bearer "+cfg.AuthToken)
+	req.Header.Set("Authorization", config.AuthHeaderPrefix+cfg.AuthToken)
 
 	client := &http.Client{Timeout: httpclient.Timeout}
 	resp, err := client.Do(req)
